requirements.txt
torch
transformers
diffusers
accelerate
protobuf
sentencepiece
flask
pillow
numpy
